refactor(cfg-builder): option to avoid boolean CrabIR statements
The option is part of CfgBuilderParams and it is called avoid_boolean.
By default, it is true so the translation will try to avoid adding
some boolean CrabIR statements such as bool_assume and
bool_assert. This is the same behavior than before so this commit does
not bring any functional change unless the flag is disabled.